Added files via upload
R code that will create an extensive PCA output.
Created in R Studio.
Based on the Orchid Output.
Lines that will have to be changed for the butterfly files to work properly: 53:54, all labels in 111:126

I'm going to have to do something about the readability of the output, Especially the correlations part is mind-bogglingly huge. But at least it's working. If it freezes up a bit  at 60% when compiling the knit-file, that's normal, don't worry. Those are the correlations.

Also my apologies if this is in the wrong repository, this is the first time I upload something to Github this way.