Add files via upload.
The file '20160411 identificaties.xlsx' is the original file with taxonomy of the second set of the Javanese butterfly collection. The file 'Butterfly_identification_by_Jan_Moonen_set2.csv' contains almost the same information, but some columns were edited for usability and texts were translated to English. This second set consists of 859 butterflies.