Add files via upload.
The file '20160626 identificaties.xlsx' is the original file with taxonomy of the third set of the Javanese butterfly collection. The file 'Butterfly_identification_by_Jan_Moonen_set3.csv' contains almost the same information, but some columns were edited for usability and texts were translated to English. This third set consists of 612 butterflies.